+++ packages/storage-plugin/src/engines.ts
@@ -1,14 +1,14 @@
-import { InjectionToken, PLATFORM_ID, inject } from '@angular/core';
-import { isPlatformBrowser } from '@angular/common';
+import { InjectionToken } from '@angular/core';
 import { StorageEngine } from '@ngxs/storage-plugin/internals';
 
 declare const ngDevMode: boolean;
+declare const ngServerMode: boolean;
 
 export const LOCAL_STORAGE_ENGINE = /* @__PURE__ */ new InjectionToken<StorageEngine | null>(
   typeof ngDevMode !== 'undefined' && ngDevMode ? 'LOCAL_STORAGE_ENGINE' : '',
   {
     providedIn: 'root',
-    factory: () => (isPlatformBrowser(inject(PLATFORM_ID)) ? localStorage : null)
+    factory: () => (typeof ngServerMode !== 'undefined' && ngServerMode ? null : localStorage)
   }
 );
 
@@ -16,6 +16,7 @@ export const SESSION_STORAGE_ENGINE = /* @__PURE__ */ new InjectionToken<Storage
   typeof ngDevMode !== 'undefined' && ngDevMode ? 'SESSION_STORAGE_ENGINE' : '',
   {
     providedIn: 'root',
-    factory: () => (isPlatformBrowser(inject(PLATFORM_ID)) ? sessionStorage : null)
+    factory: () =>
+      typeof ngServerMode !== 'undefined' && ngServerMode ? null : sessionStorage
   }
 );